refactor: separate interfaces and prune unused code
